# openapi.json
{
  "openapi": "3.0.0",
  "components": {
  "securitySchemes": {
    "BearerAuth": {
      "type": "http",
      "scheme": "bearer",
      "bearerFormat": "JWT"
    }
  }
}
  ,
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "Simple CRUD API for managing tasks."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/auth/signup": {
  "post": {
    "summary": "Register a new user",
    "tags": ["Authentication"],
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "example": "user@gmail.com"
              },
              "password": {
                "type": "string",
                "example": "Password123!"
              }
            }
          }
        }
      }
    },
    "responses": {
      "201": {
        "description": "User registered successfully"
      },
      "400": {
        "description": "Invalid request"
      }
    }
  }
}
  ,
  "/auth/login": {
  "post": {
    "summary": "Login user",
    "tags": ["Authentication"],
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "example": "user@gmail.com"
              },
              "password": {
                "type": "string",
                "example": "Password123!"
              }
            }
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Login successful"
      },
      "401": {
        "description": "Invalid credentials"
      }
    }
  }
},
"/auth/logout": {
  "post": {
    "summary": "Logout user",
    "tags": ["Authentication"],
    "responses": {
      "200": {
        "description": "Logout successful"
      }
    }
  }
},
    "/tasks": {
      "get": {
        "summary": "Get all tasks",
        "security": [
  {
    "BearerAuth": []
  }
],
        "responses": {
          "200": {
            "description": "List of tasks"
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "security": [
  {
    "BearerAuth": []
  }
],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  }
                },
                "required": ["title"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created"
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get a task by ID",
        "security": [
  {
    "BearerAuth": []
  }
],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task found"
          },
          "404": {
            "description": "Task not found"
          }
        }
      },
      "put": {
        "summary": "Update a task",
        "security": [
  {
    "BearerAuth": []
  }
],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "done": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated"
          },
          "404": {
            "description": "Task not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "security": [
  {
    "BearerAuth": []
  }
],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    }
  }
}